backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2
pydantic[email]
passlib[bcrypt]
python-multipart
python-jose